services/api-gateway/pyproject.toml
[project]
name = "ts-portal-api-gateway"
version = "0.1.0"
description = "TS Portal API Gateway - 마이크로서비스 통합 게이트웨이"
authors = [
    {name = "Seungdo Bae", email = "seungdobae@saltware.co.kr"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.19.0",
    "redis>=5.0.1",
    "aioredis>=2.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# 🔧 빌드 설정 수정 - 패키지 경로 명시
[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "app/main.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
] 